Started preview example page
The preview pages on Wacky Buttons currently use tables. This replaces
the tables with divs and bootstrap. This will serve as a template
because the preview pages are dynamically generated with php on Wacky
Buttons. Once we get into the php stage we can copy the html from this
into php.